// src/models/tVRRecord.ts
/**
 * Generated by orval
 * Do not edit manually.
 */
import type { SimpleTVRPeriod } from './simpleTVRPeriod'
import type { SimpleLegalEntity } from './simpleLegalEntity'
import type { Jurisdiction } from './jurisdiction'
import type { TaxType } from './taxType'
import type { FilingFrequency } from './filingFrequency'
import type { FilingType } from './filingType'

export interface TVRRecord {
  id?: number
  readonly period: SimpleTVRPeriod
  readonly legal_entity: SimpleLegalEntity
  readonly jurisdiction: Jurisdiction
  readonly tax_type: TaxType
  readonly filing_frequency: FilingFrequency
  readonly filing_method: FilingType
  period_id: number
  legal_entity_id: number
  jurisdiction_id: number
  /** @nullable */
  tax_type_id?: number | null
  filing_frequency_id: number
  filing_method_id: number
  /**
   * @nullable
   * @pattern ^-?\d{0,13}(?:\.\d{0,2})?$
   */
  gl_amount?: string | null
  /**
   * @nullable
   * @pattern ^-?\d{0,13}(?:\.\d{0,2})?$
   */
  sales_tax_extract?: string | null
  /**
   * @nullable
   * @pattern ^-?\d{0,13}(?:\.\d{0,2})?$
   */
  amount_to_adjust?: string | null
  /**
   * @nullable
   * @pattern ^-?\d{0,13}(?:\.\d{0,2})?$
   */
  manual_adjustment?: string | null
  /**
   * @nullable
   * @pattern ^-?\d{0,13}(?:\.\d{0,2})?$
   */
  use_tax?: string | null
  /**
   * @nullable
   * @pattern ^-?\d{0,13}(?:\.\d{0,2})?$
   */
  nmr_cf_prior?: string | null
  /**
   * @nullable
   * @pattern ^-?\d{0,13}(?:\.\d{0,2})?$
   */
  nmr_cf_future?: string | null
  /**
   * @nullable
   * @pattern ^-?\d{0,13}(?:\.\d{0,2})?$
   */
  credits_cf_prior?: string | null
  /**
   * @nullable
   * @pattern ^-?\d{0,13}(?:\.\d{0,2})?$
   */
  credits_cf_future?: string | null
  /**
   * @nullable
   * @pattern ^-?\d{0,13}(?:\.\d{0,2})?$
   */
  local_adjustment?: string | null
  /**
   * @nullable
   * @pattern ^-?\d{0,13}(?:\.\d{0,2})?$
   */
  gross_due?: string | null
  /**
   * @nullable
   * @pattern ^-?\d{0,13}(?:\.\d{0,2})?$
   */
  prepayment_credit?: string | null
  /**
   * @nullable
   * @pattern ^-?\d{0,13}(?:\.\d{0,2})?$
   */
  prepayment_due?: string | null
  /**
   * @nullable
   * @pattern ^-?\d{0,13}(?:\.\d{0,2})?$
   */
  vendors_discount?: string | null
  /**
   * @nullable
   * @pattern ^-?\d{0,13}(?:\.\d{0,2})?$
   */
  bo_tax?: string | null
  /**
   * @nullable
   * @pattern ^-?\d{0,13}(?:\.\d{0,2})?$
   */
  rounding?: string | null
  /**
   * @nullable
   * @pattern ^-?\d{0,13}(?:\.\d{0,2})?$
   */
  currency_converted?: string | null
  /** @maxLength 10 */
  currency_code?: string
  /**
   * @nullable
   * @pattern ^-?\d{0,13}(?:\.\d{0,2})?$
   */
  net_due?: string | null
  /**
   * @nullable
   * @pattern ^-?\d{0,13}(?:\.\d{0,2})?$
   */
  amount_to_fund?: string | null
  /**
   * @maxLength 255
   * @nullable
   */
  status_conf_num?: string | null
  /**
   * @maxLength 255
   * @nullable
   */
  payment_conf_num?: string | null
  /**
   * @nullable
   * @pattern ^-?\d{0,13}(?:\.\d{0,2})?$
   */
  payment_amount?: string | null
  /** @nullable */
  filing_date?: string | null
  /** @nullable */
  payment_date?: string | null
  /** @nullable */
  client_comment?: string | null
  /** @nullable */
  dstax_comment?: string | null
  is_ready?: boolean
}